feat(product image): update and create images

+++ src/main/java/com/pqkhang/ct553_backend/domain/booking/product/product_image/ProductImageServiceImpl.java
@@ -2,12 +2,12 @@ package com.pqkhang.ct553_backend.domain.booking.product.product_image;
 
 import com.pqkhang.ct553_backend.app.exception.ResourceNotFoundException;
 import com.pqkhang.ct553_backend.domain.booking.product.Product;
+import com.pqkhang.ct553_backend.domain.booking.product.ProductRepository;
 import com.pqkhang.ct553_backend.infrastructure.utils.CloudinaryUtils;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
@@ -22,6 +22,7 @@ public class ProductImageServiceImpl implements ProductImageService {
     ProductImageRepository productImageRepository;
     ProductImageMapper productImageMapper;
     CloudinaryUtils cloudinaryUtils;
+    ProductRepository productRepository;
 
     private ProductImageDTO convertAndUploadProductImageList(ProductImageDTO productImageDTO, List<MultipartFile> productImageFiles) {
         // convert product image files to list of files
@@ -34,7 +35,22 @@ public class ProductImageServiceImpl implements ProductImageService {
         }).toList();
 
         // upload images to cloudinary
-        List<String> productImageUrls = productImages.stream().map(cloudinaryUtils::uploadImage).toList();
+        String productName = productRepository.findProductByProductId(productImageDTO.getProductId()).getProductName();
+//        List<String> productImageUrls = productImages.stream().map(file -> {
+//            try {
+//                return cloudinaryUtils.uploadImage(file, productName);
+//            } catch (Exception e) {
+//                throw new RuntimeException("Error uploading image");
+//            }
+//        }).toList();
+        List<String> productImageUrls = new java.util.ArrayList<>(List.of());
+        for (int i = 0; i < productImages.size(); i++) {
+            try {
+                productImageUrls.add(cloudinaryUtils.uploadImage(productImages.get(i), productName + " - Img " + (i + 1)));
+            } catch (Exception e) {
+                throw new RuntimeException("Error uploading image");
+            }
+        }
 
         // save product images to database
         List<ProductImage> productImageList = productImageUrls.stream().map(url -> {
@@ -55,19 +71,27 @@ public class ProductImageServiceImpl implements ProductImageService {
     }
 
     @Override
-    public ProductImageDTO createProductImage(@RequestPart("productImageDTO") ProductImageDTO productImageDTO, @RequestPart("imageUrl") List<MultipartFile> productImageFiles) {
-        return convertAndUploadProductImageList(productImageDTO, productImageFiles);
+    public void createProductImage(ProductImageDTO productImageDTO, List<MultipartFile> productImageFiles) {
+        convertAndUploadProductImageList(productImageDTO, productImageFiles);
     }
 
     @Override
-    public ProductImageDTO updateAllProductImageByProductId(@RequestPart("productImageDTO") ProductImageDTO productImageDTO, @RequestPart("imageUrl") List<MultipartFile> productImageFiles) throws ResourceNotFoundException {
+    public ProductImageDTO updateAllProductImageByProductId(ProductImageDTO productImageDTO, List<MultipartFile> productImageFiles) throws ResourceNotFoundException {
+        if (productRepository.findProductByProductId(productImageDTO.getProductId()) == null) {
+            throw new ResourceNotFoundException("Product not found");
+        }
 
-        // delete all old product images
+        if (productImageFiles == null || productImageFiles.isEmpty()) {
+            return productImageDTO;
+        }
+
+        // delete all old product images in cloudinary and database
         List<ProductImage> oldProductImages = productImageRepository.findAllByProduct_ProductId(productImageDTO.getProductId());
         oldProductImages.forEach(image -> {
             try {
+                cloudinaryUtils.deleteImage(image.getImageUrl());
                 deleteProductImage(image.getProductImageId());
-            } catch (ResourceNotFoundException e) {
+            } catch (Exception e) {
                 throw new RuntimeException(e);
             }
         });
